Revert "Without amplify"
