+++ components/PostCard.tsx
@@ -37,11 +37,7 @@ export function PostCard({ post }) {
             </a>
           </p>
           <a
-            href={
-              post.published
-                ? `/posts/${post.slug}`
-                : `/api/preview?slug=${post.slug}`
-            }
+            href={`/${post.published ? "posts" : "drafts"}/${post.slug}`}
             className="block"
           >
             <h3 className="mt-2 text-xl font-semibold leading-7 text-gray-900">
