+++ package.json
@@ -16,9 +16,12 @@
     "react-markdown": "^4.3.1"
   },
   "devDependencies": {
+    "@types/node": "^14.11.10",
+    "@types/react": "^16.9.53",
     "postcss-flexbugs-fixes": "4.2.1",
     "postcss-preset-env": "^6.7.0",
-    "tailwindcss": "^1.9.4"
+    "tailwindcss": "^1.9.4",
+    "typescript": "^4.0.3"
   },
   "license": "MIT"
 }
