+++ pages/_app.tsx
@@ -1,16 +1,10 @@
 import { Footer } from "components/Footer";
-import Head from "next/head";
 import "../styles/index.css";
 
 function MyApp({ Component, pageProps }) {
   return (
     <>
-      <Head>
-        <title>Next.js + Ampify Blog</title>
-      </Head>
-
       <Component {...pageProps} />
-
       <Footer />
     </>
   );
